Differentiate between primary and secondary links

+++ src/updater/main.c
@@ -199,7 +199,9 @@ static int find_deps(struct graph *g, struct node *n)
 	struct flist f;
 	char object_dir[] = ".tup/object/" SHA1_XD;
 	char namefile[] = ".tup/object/" SHA1_XD "/.name";
+	char depfile[] = ".tup/object/" SHA1_XD "/.secondary";
 	struct stat st;
+	struct stat st2;
 
 	tupid_to_xd(object_dir + 12, n->tupid);
 	flist_foreach(&f, object_dir) {
@@ -211,11 +213,16 @@ static int find_deps(struct graph *g, struct node *n)
 			return -1;
 		}
 		tupid_to_xd(namefile + 12, f.filename);
+		tupid_to_xd(depfile + 12, f.filename);
 		if(stat(namefile, &st) < 0) {
 			perror(namefile);
 			return -1;
 		}
-		if(st.st_ino != f._ent->d_ino) {
+		if(stat(depfile, &st2) < 0) {
+			perror(depfile);
+			return -1;
+		}
+		if(f._ent->d_ino != st.st_ino && f._ent->d_ino != st2.st_ino) {
 			DEBUGP("Removing obsolete link %.*s -> %.*s\n",
 			       8, n->tupid, 8, f.filename);
 			unlinkat(f.dirfd, f.filename, 0);
